I've set up the JUCE unit test runner for DebugUIPanelTests.
This commit introduces the necessary infrastructure to run JUCE-based unit tests, specifically targeting DebugUIPanelTests.

Key changes:
- I added `test/source/TestMain.cpp` which contains the main() entry point that initializes JUCE and runs `juce::UnitTestRunner`.
- I modified `test/CMakeLists.txt`:
    - I changed the test executable target `PointilSynthTests` to be built using `juce_add_console_application`.
    - I specified `TestMain.cpp` and `DebugUIPanelTest.cpp` as sources.
    - I listed required JUCE modules (`juce_core`, `juce_events`, `juce_gui_basics`, `juce_unittest`, `juce_audio_processors`) in the `MODULES` argument of `juce_add_console_application`.
    - I linked the main plugin library (`PointillisticSynth`) to the test executable.
- I verified that `DebugUIPanelTest.cpp` includes a static instance of the test class for discovery by the `UnitTestRunner`.

Note: While these changes correctly configure the structure for a JUCE-based test executable, the `PointilSynthTests` executable itself is still failing to build. I suspect this is due to the JUCE framework (via cpm) not making the CMake targets for required modules (e.g., `juce::juce_unittest`, `juce::juce_audio_processors`) globally available for linking by standalone executables other than the main plugin. Resolving this likely requires adjustments to the root CMakeLists.txt or deeper insight into the build errors for the test target, which were not available.

The previous Google Test file (`AudioProcessorTest.cpp`) is now effectively orphaned by this setup for the `PointilSynthTests_RunTests` CTest entry.

+++ plugin/source/StochasticModel.cpp
@@ -69,7 +69,7 @@ void StochasticModel::generateNewGrain(Grain& newGrain)
         // For now, using a common default.
         currentSampleRate = 44100.0;
     }
-    newGrain.durationInSamples = static_cast<int>((randomizedDurationMs / 1000.0) * currentSampleRate);
+    newGrain.durationInSamples = static_cast<int>((static_cast<double>(randomizedDurationMs) / 1000.0) * currentSampleRate);
 
     // Other members of Grain (pitch, pan, amplitude, etc.) are not set here
     // as per the current task. They would be set by other parts of StochasticModel
@@ -110,7 +110,7 @@ int StochasticModel::getSamplesUntilNextEvent()
         return INT_MAX;
     }
 
-    double averageSamplesPerGrain = currentSampleRate / currentGrainsPerSecond;
+    double averageSamplesPerGrain = currentSampleRate / static_cast<double>(currentGrainsPerSecond);
 
     // Check for potential issues with averageSamplesPerGrain before using it,
     // especially with the Poisson distribution.
